+++ packages/cli/src/projects/options.ts
@@ -9,6 +9,8 @@ export type Opts = BaseOpts & {
   removeUnmapped?: boolean | undefined;
   workflowMappings?: Record<string, string> | undefined;
   project?: string;
+  format?: 'yaml' | 'json' | 'state';
+  clean?: boolean;
 };
 
 // project specific options
@@ -28,11 +30,29 @@ export const alias: CLIOption = {
   },
 };
 
+export const clean: CLIOption = {
+  name: 'clean',
+  yargs: {
+    description: 'Clean the working dir before checking out the new project',
+    default: false,
+    boolean: true,
+  },
+};
+
 export const dryRun: CLIOption = {
   name: 'dryRun',
   yargs: {
     description:
       'Runs the command but does not commit any changes to disk or app',
+  },
+};
+
+export const format: CLIOption = {
+  name: 'format',
+  yargs: {
+    hidden: true,
+    description:
+      'The format to save the project as - state, yaml or json. Use this to download raw state files.',
   },
 };