+++ packages/cli/test/projects/deploy.test.ts
@@ -1,9 +1,44 @@
+import { readFile, writeFile } from 'node:fs/promises';
 import test from 'ava';
+import mock from 'mock-fs';
+import path from 'node:path';
 import Project, { generateWorkflow } from '@openfn/project';
 import { createMockLogger } from '@openfn/logger';
-import { reportDiff } from '../../src/projects/deploy';
+import createLightningServer, {
+  DEFAULT_PROJECT_ID,
+} from '@openfn/lightning-mock';
+
+import {
+  handler as deployHandler,
+  reportDiff,
+} from '../../src/projects/deploy';
+import { myProject_yaml, myProject_v1 } from './fixtures';
+import { checkout } from '../../src/projects';
 
 const logger = createMockLogger(undefined, { level: 'debug' });
+
+const port = 9876;
+const ENDPOINT = `http://localhost:${port}`;
+
+let server: any;
+
+test.before(async () => {
+  server = await createLightningServer({ port });
+});
+
+test.beforeEach(() => {
+  server.addProject(myProject_v1);
+  logger._reset();
+  mock.restore();
+});
+
+const mockFs = (paths: Record<string, string>) => {
+  const pnpm = path.resolve('../../node_modules/.pnpm');
+  mock({
+    [pnpm]: mock.load(pnpm, {}),
+    ...paths,
+  });
+};
 
 // what will deploy tests look like?
 
@@ -148,3 +183,98 @@ test('reportDiff: should report mix of added, changed, and removed workflows', (
   t.truthy(logger._find('always', /workflows removed/i));
   t.truthy(logger._find('always', /- c/i));
 });
+
+// This doesn't work until local history is tracked properly
+test.serial.skip(
+  'deploy a change to a project and write the yaml back (compatible histories)',
+  async (t) => {
+    mockFs({
+      '/ws/.projects/main@app.openfn.org.yaml': myProject_yaml,
+      '/ws/openfn.yaml': '',
+    });
+
+    // first checkout the project
+    await checkout(
+      {
+        project: 'main',
+        workspace: '/ws',
+      },
+      logger
+    );
+
+    // Now change the expression
+    await writeFile('/ws/workflows/my-workflow/transform-data.js', 'log()');
+
+    await deployHandler(
+      {
+        endpoint: ENDPOINT,
+        apiKey: 'test-api-key',
+        workspace: '/ws',
+        log: 'debug',
+      } as any,
+      logger
+    );
+
+    // Check what was uploaded to Lightning - the internal app state
+    // should be the exact state object that was uploaded
+    const uploadedState =
+      server.state.projects['e16c5f09-f0cb-4ba7-a4c2-73fcb2f29d00'];
+    t.truthy(uploadedState);
+
+    const projectYaml = await readFile(
+      '/ws/.projects/main@app.openfn.org.yaml',
+      'utf8'
+    );
+    t.regex(projectYaml, /fn()/);
+
+    const success = logger._find('success', /Updated project at/);
+    t.truthy(success);
+  }
+);
+
+// TODO skipping while history checking is messed up
+test.serial.skip(
+  'Exit early if the remote is not compatible with local',
+  async (t) => {
+    mockFs({
+      '/ws/.projects/main@app.openfn.org.yaml': myProject_yaml,
+      '/ws/openfn.yaml': '',
+    });
+
+    // Update the server-side project
+    const changed = JSON.parse(JSON.stringify(myProject_v1));
+    changed.workflows['my-workflow'].version_history.push('app:abc');
+
+    server.addProject('e16c5f09-f0cb-4ba7-a4c2-73fcb2f29d00', changed);
+
+    // checkout the project locally
+    await checkout(
+      {
+        project: 'main',
+        workspace: '/ws',
+      },
+      logger
+    );
+
+    // Now change the expression
+    await writeFile('/ws/workflows/my-workflow/transform-data.js', 'log()');
+
+    await deployHandler(
+      {
+        endpoint: ENDPOINT,
+        apiKey: 'test-api-key',
+        workspace: '/ws',
+      } as any,
+      logger
+    );
+
+    // The remote project should not have changed
+    const appState =
+      server.state.projects['e16c5f09-f0cb-4ba7-a4c2-73fcb2f29d00'];
+    t.deepEqual(appState, myProject_v1);
+
+    // We should log what's going on to the user
+    const expectedLog = logger._find('error', /projects have diverged/i);
+    t.truthy(expectedLog);
+  }
+);